started renaming web apps and splitting functionality
